src/main/java/at/jku/dke/task_app/xquery/data/entities/XQuerySubmission.java
package at.jku.dke.task_app.xquery.data.entities;

import at.jku.dke.etutor.task_app.data.entities.BaseSubmission;
import at.jku.dke.etutor.task_app.dto.SubmissionMode;
import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.Table;
import jakarta.validation.constraints.NotNull;

/**
 * Represents a XQuery input.
 */
@Entity
@Table(name = "submission")
public class XQuerySubmission extends BaseSubmission<XQueryTask> {
    @NotNull
    @Column(name = "submission", nullable = false)
    private String submission;

    /**
     * Creates a new instance of class {@link XQuerySubmission}.
     */
    public XQuerySubmission() {
    }

    /**
     * Creates a new instance of class {@link XQuerySubmission}.
     *
     * @param submission The input.
     */
    public XQuerySubmission(String submission) {
        this.submission = submission;
    }

    /**
     * Creates a new instance of class {@link XQuerySubmission}.
     *
     * @param userId        The user id.
     * @param assignmentId  The assignment id.
     * @param task          The task.
     * @param language      The language.
     * @param feedbackLevel The feedback level.
     * @param mode          The mode.
     * @param submission    The input.
     */
    public XQuerySubmission(String userId, String assignmentId, XQueryTask task, String language, int feedbackLevel, SubmissionMode mode, String submission) {
        super(userId, assignmentId, task, language, feedbackLevel, mode);
        this.submission = submission;
    }

    /**
     * Gets the input.
     *
     * @return The input.
     */
    public String getSubmission() {
        return submission;
    }

    /**
     * Sets the input.
     *
     * @param submission The input.
     */
    public void setSubmission(String submission) {
        this.submission = submission;
    }
}
